+++ src/gitbook_downloader/gui/web/assets/swimlanesDiagram-VR7AAH4N-B2UJfnQ7.js
@@ -1,0 +1,8 @@
+import{c as e,s as r}from"./flowDiagram-HODETNUW-B-rP7RcB.js";import{_ as a}from"./index-BQslKoQ7.js";import"./chunk-5VM5RSS4-CaeJmjpG.js";import"./chunk-XXDRQBXY-C_gUF7g1.js";import"./chunk-POPQ4Y6H-CbAbmnqT.js";import"./chunk-F27PBJKO-CeAOf_RV.js";import"./channel-XMgEjEiR.js";var o=a(t=>`${r(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,f=e({defaultLayout:"swimlane",styles:s});export{f as diagram};